requirements.txt
PyQt6==6.6.1
PyQt6-WebEngine==6.6.1
PyYAML==6.0
pytest==7.4.3
pytest-cov==4.1.0